+++ Tools/rdmp/CommandLine/Gui/ConsoleGuiOptions.cs
@@ -16,5 +16,10 @@ namespace Rdmp.Core.CommandLine.Gui
     public class ConsoleGuiOptions : RDMPCommandLineOptions
     {
 
+        /// <summary>
+        /// Sets UseSystemConsole to true for Terminal.gui (i.e. uses the NetDriver which is based on System.Console)
+        /// </summary>
+        [Option("usc", HelpText = "Sets UseSystemConsole to true for Terminal.gui (i.e. uses the NetDriver which is based on System.Console)")]
+        public bool UseSystemConsole { get; internal set; }
     }
 }